Update Vortice.Vulkan package and cast index length
Upgraded Vortice.Vulkan package from version 1.7.1 to 1.8.4 in the csproj file. Added explicit cast to uint for _graphicsDevice.Indices.Length in TestApp.cs for better type consistency.

+++ VulkanTest/TestApp.cs
@@ -62,7 +62,7 @@ public unsafe class TestApp : Application
         vkCmdBindIndexBuffer(commandBuffer, _graphicsDevice.IndexBuffer, 0, VkIndexType.Uint16);
 
         //vkCmdDraw(commandBuffer, (uint)_graphicsDevice.Vertices.Length, 1, 0, 0);
-        vkCmdDrawIndexed(commandBuffer, _graphicsDevice.Indices.Length, 1, 0, 0, 0);
+        vkCmdDrawIndexed(commandBuffer, (uint)_graphicsDevice.Indices.Length, 1, 0, 0, 0);
 
         vkCmdSetBlendConstants(commandBuffer, new Vector4(1.0f, 1.0f, 1.0f, 1.0f));
         vkCmdEndRenderPass(commandBuffer);
